Separate changes from EMV app
