+++ lib/nfc/protocols/emv.h
@@ -32,11 +32,6 @@ typedef struct {
 } EmvData;
 
 typedef struct {
-    uint8_t aid[16];
-    uint16_t aid_len;
-} IdData;
-
-typedef struct {
     uint16_t tag;
     uint8_t data[];
 } PDOLValue;
@@ -62,11 +57,6 @@ typedef struct {
     APDU afl;
 } EmvApplication;
 
-typedef struct {
-    uint8_t aid[16];
-    uint8_t aid_len;
-} IdApplication;
-
 /** Read bank card data
  * @note Search EMV Application, start it, try to read AID, PAN, card name,
  * expiration date, currency and country codes
@@ -77,8 +67,6 @@ typedef struct {
  * @return true on success
  */
 bool emv_read_bank_card(FuriHalNfcTxRxContext* tx_rx, EmvApplication* emv_app);
-
-bool read_id_card(FuriHalNfcTxRxContext* tx_rx, IdApplication* id_app);
 
 /** Search for EMV Application
  *
